Module bi_sql_excel_reports: Changed the Add-in download wizard so users can indicate which operating system they use (Windows or Apple Mac) to get tailored download files and instructions. Separated the authorization logic into a separate file. Updated the Excel Add-in to support usage of a Mac Script for users on a Mac (because of 'sandbox' restrictions).
