+++ bi_sql_excel_reports/models/__init__.py
@@ -2,7 +2,6 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 from . import bi_sql_excel_report
+from . import bi_sql_excel_report_auth
 from . import bi_sql_excel_report_field
 from . import bi_sql_view
-
-
